+++ demo/src/main/java/raven/modal/demo/Demo.java
@@ -14,7 +14,7 @@ import java.awt.*;
 
 public class Demo extends JFrame {
 
-    public static final String DEMO_VERSION = "2.4.1-SNAPSHOT";
+    public static final String DEMO_VERSION = "2.5.0";
 
     public Demo() {
         init();