+++ CHANGELOG.md
@@ -10,6 +10,11 @@
     - Add new `DrawerNoneLineStyle`. use this for no-paint drawer line style but paint arrow style
 - Other
     - Windows 11: heavyWeight windows now support rond border with drop shadow using flatlaf native library (PR #33)
+- Extras
+    - AvatarIcon: add method to change icon:
+        - `void setIcon(String filename);`
+        - `void setIcon(URL location);`
+        - `void setIcon(Icon icon);`
 
 ### Changed